Clean up

+++ install_osx_software.sh
@@ -3,8 +3,22 @@
 # Install software for the Mac OSX
 # This is called from the `install.sh` script
 #
+
+# Homebrew needs to be installed first
+which brew > /dev/null;
+if [[ $? -neq 0 ]]; then
+	echo "Please install homebrew first. See https://brew.sh"; 
+	exit 1
+fi
+
+# Miniconda3 needs to be installed first
+if [[ ! -e ${HOME}/miniconda3 ]]; then
+	echo "Please install Miniconda3 first. See https://conda.io/miniconda.html";
+	exit 1;
+fi
+
 brew install wget 
-bew install neovim 
+brew install neovim 
 
 # Fancy top command
 brew install htop
@@ -23,6 +37,18 @@ brew install grip
 # https://apple.stackexchange.com/a/209269
 brew install bash-completion
 
+# A code-searching tool similar to ack, but faster
+brew install the_silver_searcher
+
+#----------
+# Tmux
+#----------
+
+brew install tmux 
+
+# Allow for copying from inside a tmux session using vim commands
+brew install reattach-to-user-namespace --with-wrap-pbcopy-and-pbpaste
+
 #----------
 # fzf
 #----------
@@ -30,12 +56,11 @@ brew install fzf
 
 # To install useful key bindings and fuzzy completion:
 $(brew --prefix)/opt/fzf/install \
-	--key-binding \
+	--key-bindings \
 	--completion
-
-#----------
-# Pygments
-#----------
 
 # Used for syntax highlighting
 conda install -y Pygments
+
+# conda tab complete
+conda install argcomplete
